Add markers and fix setup issues

+++ lib/feature/domain/entities/location_entity.dart
@@ -1,6 +1,6 @@
 class LocationEntity {
-  final int lat;
-  final int lng;
+  final double lat;
+  final double lng;
 
   LocationEntity({required this.lat, required this.lng});
 }
